test(): move tests to separate folder
