+++ .projections.json
@@ -1,7 +1,7 @@
 {
-  "src/*.rs": { "type": "source" },
-  "benches/*.rs": { "type": "bench", "alternate": "tests/{}.rs" },
+  "src/*.rs": { "type": "source", "alternate": "tests/{}.rs" },
   "tests/*.rs": { "type": "test", "alternate": "benches/{}.rs" },
+  "benches/*.rs": { "type": "bench", "alternate": "src/{}.rs" },
   "README.md": { "type": "doc" },
   "docs/*": { "type": "doc" }
 }
