+++ Cargo.toml
@@ -41,7 +41,7 @@ block = ["blowfish"]
 blowfish = []
 chacha20 = []
 crypto = ["symmetric", "public-key"]
-default = ["digest", "crypto", "ssl", "kdf"]
+default = ["digest", "crypto", "ssl", "kdf", "mac"]
 digest = ["digest-deprecated", "digest-safe"]
 digest-deprecated = ["md4", "md5", "ripemd", "sha1", "tiger"]
 digest-safe = ["sha2", "sha3"]
