+++ src/mac/mod.rs
@@ -1,4 +1,5 @@
-#[cfg(feature = "hmac")] pub mod hmac;
+#[cfg(feature = "hmac")]
+pub mod hmac;
 
 pub trait Mac: Sized {
     fn update<D: AsRef<[u8]>>(&mut self, data: D);
