+++ tests/crypto/block/mod.rs
@@ -1,0 +1,1 @@
+mod blowfish;
