+++ tests/crypto/mod.rs
@@ -1,0 +1,2 @@
+mod block;
+mod stream;
