+++ tests/crypto/stream/mod.rs
@@ -1,0 +1,1 @@
+mod chacha20;
