+++ tests/lib.rs
@@ -1,0 +1,8 @@
+extern crate octavo;
+extern crate openssl;
+extern crate quickcheck;
+
+mod crypto;
+mod digest;
+mod kdf;
+mod mac;
