+++ tests/mac/mod.rs
@@ -1,0 +1,1 @@
+mod hmac;
